Move collector and stream into Util
