+++ packages/punchcard/lib/core/index.ts
@@ -1,9 +1,7 @@
 export * from './assembly';
 export * from './cache';
 export * from './client';
-export * from './collector';
 export * from './dependency';
 export * from './entrypoint';
 export * from './file-system';
 export * from './resource';
-export * from './stream';
