+++ packages/punchcard/lib/kinesis/stream.ts
@@ -11,12 +11,12 @@ import { Cache } from '../core/cache';
 import { Clients } from '../core/client';
 import { Dependency } from '../core/dependency';
 import { Resource } from '../core/resource';
-import { Stream as SStream } from '../core/stream';
 import { DeliveryStream } from '../firehose/delivery-stream';
 import { BufferMapper, Json, Mapper, RuntimeType, Type } from '../shape';
 import { Codec } from '../util/codec';
 import { Compression } from '../util/compression';
 import { sink, Sink, SinkProps } from '../util/sink';
+import { Stream as SStream } from '../util/stream';
 import { Event } from './event';
 
 export type Config = SStream.Config & events.KinesisEventSourceProps;
