+++ packages/punchcard/lib/util/index.ts
@@ -1,7 +1,9 @@
 export * from './codec';
+export * from './collector';
 export * from './compression';
 export * from './constants';
 export * from './hlist';
 export * from './sink';
+export * from './stream';
 export * from './tree';
-export * from './utils';
+export * from './omit';
